- Fixed /iris remove for real this time lol - Changed it back to /iris instead of /iris worldmanager for easier access
